+++ js/modules/default-modal.js
@@ -164,6 +164,10 @@ $(function(){
             // init
             var $trigger = $(this);
             var initialized_attr = 'data-trigger-initialized';
+            // trigger not visible? Adios!
+            if($trigger.is(':visible') === false) {
+                return true;
+            }
             // check for initialized trigger
             var trigger_initialized = $trigger.attr(initialized_attr);
             // NOT initialized yet