+++ js/modules/image-modal.js
@@ -128,7 +128,7 @@ $(function(){
             var initialized_attr = 'data-trigger-initialized';
             // trigger not visible? Adios!
             if($trigger.is(':visible') === false) {
-                return false;
+                return true;
             }
             // check for initialized trigger
             var trigger_initialized = $trigger.attr(initialized_attr);