+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Each release is divided into the following main categories:
   $newsletter-form-modal-max-width-xxl:        550px;
 ```
 ### NEW
+- Swiper now supports instance specific transition durations between slides that can be set on the `swiper-container` via `data`-attributes.
 - Google Map now supports location specific pins by setting a `pin_url` (besides `lat`, `lng` and `infowindow_content`).
 - Video poster and source are now lazyloaded.
 - `softpage` and `default-modal` now support a `tingle-modal-header`. Details in both javascript files.